Add contact section and footer
I have started working on the contact section but having some problems
with the alignment between the icons and the text info next to it. I
will try to figure it out next time working on the project. I also
added a footer with the copyright and still need to add the social
networks icons.