Redo the social media icons
I have removed the social media icons in the contact section and have
created new ones from scratch to have white background circles on the
black footer. All icons are png format and the logo for each social
platform is showing the background colour the icon is on, in this case
the black colour. All icons will open a new tab.