Add media queries first version
I am trying to make the website responsive so I am adding max-width
because have started with big screen. I have added two breakpoints, not
sure if it is correct or not to do it that way but I am focusing on
trying to make it responsive as much as I could. I will have to fix
that with work section mostly and also footer plus the navigation.