Add media queries for mobile first version
I have done few steps in this max-width for mobile portrait related to
header and navigation but still have a lot to do to finish this media
query for mobile portrait. Not sure what I am doing is right or wrong.
I am going to ask for help next workshop.