Edit html and css plus start with mobile first
This was during my workshop with Jo. I removed all the ids in the HTML
except the ids linked to the navigation and replaced them with classes
and also added classes to my sections linked to navigation. Also
replaced some of the divs with HTML5 elements; header, section and
footer. For the CSS added box-sizing, added some flexboxes to some of
the selectors. Also removed my old media queries, have started with
mobile first and added the min width to have the styles that I have now
for tablets and will continue to adjust the rest of the mobile first
next time.