Fix the social media after box sizing
I have added box-sizing and I found the social media icons are
displayed differently so I fixed it but then I wanted to add display
flex to the copyright and the icons in the footer but everything is
acting weird, I think something is wrong with the heights of the footer
as the black background and the heights of the elements in that
container. Will try to figure it out next time.