Fix social media icons in desktop and mobile
I have managed to fix the display of social media icons after changing
it to flexbox in my original codes (which was starting with desktop)
then I have changed it again in mobile first design and add the desktop
design in my min width media queries. Also changed most of the class
names instead of camelCase to hyphen.