Fix media queries after workshop version one
I have done the workshop and learned new things about flexbox like
order and heights. I have done all the media queries but it just needs
few edits which I am going to work on next time. Also I have added a js
file to add an eventListener to the window and have a background colour
when the scroll reaches the header part. Still working on the script.