Fix media queries version after workshop version two
I have tried to delete all the codes I don’t need and make it DRY as
much as I could. I have made the margins and paddings in the header and
footer to be nearly the same at least visually. Still have one more
thing to do which is add padding between projects for the min width
767px. Then start the js for the header. My final stage is to add the
real content for the about, contact sections.