Add link about nav to section version one
This was during my code workshop with Jo. I was working on linking the
navigation link about with the section and have the correct height. I
am still working on it and also on the way the work section background
color isn’t working the way it supposed to be. Still need to work on it.