Fix about and work problem with padding
The problem I was experiencing with small screens when I click on the
about or work link in the navigation some of the sections content were
hidden under the header background color. I tried to do it in the
workshop last time but the solution we came to wasn’t working all the
time. I thought about why not add some padding before the content of
each section so when I click this amount of whitespace will be hidden
under the header and the beginning of the content as text or image will
appear just after the header. It worked and it served its purpose but I
believe there might be a way to do it with js. I will try to find out
next workshop what could be done.